requirements.txt
discord.py==2.3.2
aiohttp==3.8.5
python-dotenv==1.0.0
Pillow==10.2.0